Notebooks v1.0.0-a7/Keras/requirements.txt
keras
tensorflow
scikit-learn
numpy
Pillow
bentoml==1.0.0-a7